backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<24.0
beautifulsoup4==4.9.3
requests==2.32.4
google-auth>=2.35,<3.0
playwright
Pillow
